Update navigation to use relative URLs rather than absolute.
Github Pages serves this site under a folder, which causes the
navigation to break. As a quick fix, I’m switching the navigation to
use relative URLs. A better long-term solution would probably be to
change how the site is served in the development environment, so that
it is also served from the same folder-name as it is on Github Pages.